fix: use BASH_SOURCE to find the root of project
When calling local-dev-cluster/main.sh from other (git) projects, using
`git` to find the project root causes the PROJECT_ROOT to be assigned to
the git checkout directory of the caller instead of the directory where
local-dev-cluster is cloned. This is fixed by using BASH_SOURCE to
determine the root of the local-dev-cluster project.

The change uses the behavior explained above to allow caller to define a
`.env` file to override the defaults set in main.sh

Signed-off-by: Sunil Thaha <sthaha@redhat.com>